Fix xargo breaking tests, add test framework for Port
The newer versions of xargo were breaking the tests being run, so we
switched over to just using `cargo` for testing. Potentially we can
switch over to cargo for cross compiling now that our target is
supported, but we'll need to do a little research to see how that works.

The testing framework should make it much easier to test out registers,
any additional functionality that may be needed can be added fairly
easily from the test module as well.