Update test framework for port layer
Since the register exports its offset as part of its trait, we can just
use that from within the test framework instead of passing in the
offset. This should make for a simpler interface with less chance for
user error.

Also got workspaces working, so we don't have to do the workaround of
compiling the `cortex_m0` crate seperately for tests. This should make
for faster iteration times when writing tests as we won't have to
recompile dependencies.