Fix Modulo division
The compiler builtin function being called for modulo division
`__aeabi_uidivmod` uses a special calling convention that was not being
adhered to. In order to get around this we must implement the function
in assembly to match the requirements of the calling convention.
The issue was discovered from:
https://github.com/rust-lang-nursery/compiler-builtins/blob/master/src/arm.rs
(See the note at the top of the file)

This fixes #28 and #53