Update `volatile` crate for more ergonomic use
This alters the `volatile` crate to try and make it behave exactly like
raw pointers. Dereferencing the pointer should be treated as a volatile
operation by the compiler and force a load or store of the value from
memory.

These changes remove the `RawVolatile` backing type that was being used
to implement basic operations for the pointed at value. Now the `Deref`
and `DerefMut` traits of the `Volatile` type just directly dereference
the stored pointer. The volatile behavior is achieved through a little
assembly trick, inserting a

```rust
asm!("" ::: "memory" : "volatile");
```

macro tricks the compilier into believing that some assembly
instructions were generated that touched main memory, and so any loads
and stores after that must reach out to main memory again. This isn't
neccessarily ideal since it could inhibit some compiler optimizations
that would otherwise be fine, but the ease of use it provides seems like
a much greater benefit.